requirements.txt
asgiref==3.3.4
cffi==1.14.5
Django==3.2.4
milksnake==0.1.5
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
urlquote==1.1.4
Flask==2.0.3
python-dotenv==0.21.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.6
Werkzeug==2.2.3
